fix: tests, split tests into suite
