+++ packages/round-manager/src/features/round/__tests__/RoundDetailForm.test.tsx
@@ -396,101 +396,3 @@ describe("<RoundDetailForm />", () => {
     expect(screen.getByTestId("chain-logo")).toBeInTheDocument();
   });
 });
-
-describe("<QuadraticFundingForm />", () => {
-  beforeEach(() => {
-    renderWrapped(<QuadraticFundingForm stepper={FormStepper} />);
-  });
-
-  it("selects payout token", function () {
-    const payoutTokenSelection = screen.getByTestId("payout-token-select");
-    fireEvent.click(payoutTokenSelection);
-    const firstTokenOption = screen.getAllByTestId("payout-token-option")[0];
-    fireEvent.click(firstTokenOption);
-    fireEvent.change(screen.getByTestId("matching-funds-available"), {
-      target: { value: 1 },
-    });
-  });
-
-  it("renders the quadratic funding settings section", () => {
-    expect(screen.getByText(/Quadratic Funding Settings/i)).toBeInTheDocument();
-  });
-
-  it("renders a dropdown list of tokens when payout token input is clicked", async () => {
-    const options = getPayoutTokenOptions(ChainId.GOERLI_CHAIN_ID);
-
-    const payoutTokenSelection = screen.getByTestId("payout-token-select");
-    fireEvent.click(payoutTokenSelection);
-
-    const selectOptions = await screen.findAllByTestId("payout-token-option");
-    expect(selectOptions).toHaveLength(options.length);
-  });
-
-  it("shows validation error message when the payout token is not selected", async () => {
-    const payoutTokenSelection = screen.getByTestId("payout-token-select");
-    fireEvent.click(payoutTokenSelection);
-
-    fireEvent.click(screen.getByText("Launch"));
-
-    const errors = await screen.findByText(
-      "You must select a payout token for your round."
-    );
-    expect(errors).toBeInTheDocument();
-  });
-
-  it("renders matching funds available & matching cap input fields", () => {
-    expect(screen.getByTestId("matching-funds-available")).toBeInTheDocument();
-    expect(screen.getByTestId("matching-cap-selection")).toBeInTheDocument();
-    expect(screen.getByTestId("matching-cap-true")).toBeInTheDocument();
-    expect(screen.getByTestId("matching-cap-false")).toBeInTheDocument();
-    expect(screen.getByTestId("matching-cap-percent")).toBeInTheDocument();
-  });
-
-  it("enables matching cap when matching cap is selected to 'Yes'", () => {
-    fireEvent.click(screen.getByTestId("matching-cap-true"));
-
-    expect(screen.getByTestId("matching-cap-percent")).toBeEnabled();
-  });
-
-  it("defaults matching cap to be disabled", () => {
-    expect(screen.getByTestId("matching-cap-percent")).toBeDisabled();
-  });
-
-  it("shows validation error message when matching funds amount is not provided", async () => {
-    fireEvent.change(screen.getByTestId("matching-funds-available"), {
-      target: { value: "" },
-    });
-    fireEvent.click(screen.getByText("Launch"));
-
-    const errors = await screen.findByText(
-      "Matching funds available must be a valid number."
-    );
-    expect(errors).toBeInTheDocument();
-  });
-
-  it("shows validation error message when matching funds amount is <= zero", async () => {
-    fireEvent.change(screen.getByTestId("matching-funds-available"), {
-      target: { value: 0 },
-    });
-    fireEvent.click(screen.getByText("Launch"));
-
-    const errors = await screen.findByText(
-      "Matching funds available must be more than zero."
-    );
-    expect(errors).toBeInTheDocument();
-  });
-
-  it("shows validation error message if matching cap is selected and percentage is <= zero", async () => {
-    fireEvent.click(screen.getByTestId("matching-cap-true"));
-    fireEvent.change(screen.getByTestId("matching-cap-percent"), {
-      target: { value: 0 },
-    });
-    fireEvent.click(screen.getByText("Launch"));
-    console.log(prettyDOM(screen.getByTestId("matching-cap-percent")));
-
-    const errors = await screen.findByText(
-      "Matching cap amount must be more than zero."
-    );
-    expect(errors).toBeInTheDocument();
-  });
-});
